🩹 Simple Fix : 도수가 소수점도 적용되게 바꿈

+++ crawler-basic/JavaWebCrawling/src/main/java/com/sinse/javawebcrawling/domain/Product.java
@@ -28,7 +28,7 @@ public class Product {
     private String content;
     //도수
     @JsonProperty("alcohol")
-    private int alcohol;
+    private double alcohol;
     //용량
     @JsonProperty("volume")
     private int volume;
